Created BaseTestCase class for unit testing.
